+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [2.1.0]
 
 ## Added
 
@@ -45,5 +45,6 @@ Object-C, which shows (nicer) Terminal icon.
 * `-f` flag for OS X. This caused unexpected behavior for people who use iTerm2.
 * OS X-specific flags and usage text from Linux and FreeBSD help.
 
-[Unreleased]: https://github.com/variadico/noti/compare/v2.0.0...dev
+[Unreleased]: https://github.com/variadico/noti/compare/v2.1.0...dev
+[2.1.0]: https://github.com/variadico/noti/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/variadico/noti/compare/v1.3.0...v2.0.0